feat: set final figures for VIP-226 (BUSD sunset)
